@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build"
+    "build": "react-scripts build",
+    "start": "react-scripts start"
   },
   "browserslist": {
     "production": [
@@ -25,16 +25,21 @@
     ]
   },
   "dependencies": {
-    "@date-io/luxon": "^1.3.13",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/pickers": "^3.3.10",
-    "luxon": "^1.27.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@date-io/luxon": "^2.16.1",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@fontsource/roboto": "^4.5.8",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/lab": "^5.0.0-alpha.114",
+    "@mui/material": "^5.11.0",
+    "@mui/styles": "^5.11.2",
+    "@mui/x-date-pickers": "^5.0.12",
+    "luxon": "^3.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "react-scripts": "^4.0.3",
-    "webpack": "^4.44.2"
+    "react-scripts": "^5.0.1",
+    "webpack": "^5.75.0"
   }
 }
