@@ -1,11 +1,12 @@
 import React from "react";
-import {Box, Divider, Grid, Popover, Tooltip, Typography} from "@material-ui/core";
-import {Info} from "@material-ui/icons";
 
-import {styles} from "../../styles/form.styles";
+import {Box, Divider, Grid, Popover, Tooltip, Typography} from "@mui/material";
+import {Info} from "@mui/icons-material";
+
+import {styles} from "../../../styles/form.styles";
 
 export default function FormHeading({id, text, info}) {
-    const classes = styles()
+    const {heading, icon, icons, popover} = styles()
 
     const [anchor, setAnchor] = React.useState(null)
 
@@ -37,18 +38,18 @@ export default function FormHeading({id, text, info}) {
 
     return (
         <Grid item sm={12}>
-            <Typography className={classes.heading} variant={"h6"} gutterBottom>
+            <Typography className={heading} variant={"h6"} gutterBottom>
                 {text}
-                <Box className={classes.icons} component="span">
+                <Box className={icons} component="span">
                     {info && (
                         <Tooltip key={`${id}-info`} title='Info' placement={"top"} onClick={handleOpen} arrow>
-                            <Info className={classes.icon}/>
+                            <Info className={icon}/>
                         </Tooltip>
                     )}
                     {info && (
                         <Popover key={`${info}-popover`} {...popoverProps}>
                             {info.split('\n').map((fragment, idx) =>
-                                <Typography key={`${id}-frag-${idx}`} className={classes.popover}>
+                                <Typography key={`${id}-frag-${idx}`} className={popover}>
                                     {fragment}
                                 </Typography>)}
                         </Popover>
